Upload of the second version of the j notebook
I've learned a bit more and discover that i have a concept error with my prior analysis, using an anova test for searching relationship within categorical variables. The error was using that test and the value of "p" to decide if there was a correlation within those variables. Now i corrected that using a "pointbiserialr" method that was more aligned with our target variable.